Refactored Card component to compound component
