+++ app/javascript/src/components/Dashboard/Notes/Card/constants.js
